do not print stack frame when train process killed (#4346)
* do not print stack frame when train process killed
* add note about VOT does not support windows platform!
