+++ PaddleCV/tracking/ltr/data/loader.py
@@ -1,25 +1,8 @@
 import os
-import signal
 import sys
 
 import dataflow as df
 import numpy as np
-
-
-# handle terminate reader process, do not print stack frame
-def _reader_quit(signum, frame):
-    print("Reader process exit.")
-    sys.exit()
-
-
-def _term_group(sig_num, frame):
-    print('pid {} terminated, terminate group '
-          '{}...'.format(os.getpid(), os.getpgrp()))
-    os.killpg(os.getpgid(os.getpid()), signal.SIGKILL)
-
-
-signal.signal(signal.SIGTERM, _reader_quit)
-signal.signal(signal.SIGINT, _term_group)
 
 
 class LTRLoader(df.DataFlow):
